chore: 规范代码 - 使用u_int64_t替换__uint64_t - 将必要的指针限定为cons - 将必要的int替换为size_t和ssize_t

+++ src/tui/date.c
@@ -32,6 +32,6 @@ date_redraw(const char *date_str) {
 }
 
 void
-date_setpos(int x, int y) {
+date_setpos(size_t x, size_t y) {
     setpos(date_offset, x, y);
 }
