+++ src/tui/date.h
@@ -3,6 +3,6 @@
 
 const char *date_fmt(char str[CLOCK_DATE_LEN]);
 void date_redraw(const char *date_str);
-void date_setpos(int x, int y);
+void date_setpos(size_t x, size_t y);
 
 #endif /* DATE_H */
