+++ src/tui/info.h
@@ -13,7 +13,7 @@ typedef struct {
 int info_init(info_t *info);
 int info_fetch(info_t *info);
 
-void info_setpos(int x, int y);
+void info_setpos(size_t x, size_t y);
 void info_redraw(info_t *info);
 
 #endif /* INFO_H */
